+++ server/src/main/java/com/konkuk/soar/member/repository/MemberRepository.java
@@ -1,0 +1,8 @@
+package com.konkuk.soar.member.repository;
+
+import com.konkuk.soar.member.domain.Member;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface MemberRepository extends JpaRepository<Member, Long> {
+
+}
