+++ server/src/main/java/com/konkuk/soar/portfolio/domain/id/PortfolioAndTagId.java
@@ -1,0 +1,15 @@
+package com.konkuk.soar.portfolio.domain.id;
+
+import com.konkuk.soar.common.domain.Tag;
+import com.konkuk.soar.portfolio.domain.Portfolio;
+import lombok.EqualsAndHashCode;
+import lombok.Getter;
+
+import java.io.Serializable;
+
+@Getter
+@EqualsAndHashCode
+public class PortfolioAndTagId implements Serializable {
+    private Portfolio portfolio;
+    private Tag tag;
+}
