+++ server/src/main/java/com/konkuk/soar/portfolio/repository/PortfolioRepository.java
@@ -1,0 +1,8 @@
+package com.konkuk.soar.portfolio.repository;
+
+import com.konkuk.soar.portfolio.domain.Portfolio;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface PortfolioRepository extends JpaRepository<Portfolio, Long> {
+
+}
