+++ server/src/main/java/com/konkuk/soar/portfolio/repository/PortfolioTagRepository.java
@@ -1,0 +1,9 @@
+package com.konkuk.soar.portfolio.repository;
+
+import com.konkuk.soar.portfolio.domain.PortfolioTag;
+import com.konkuk.soar.portfolio.domain.id.PortfolioAndTagId;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface PortfolioTagRepository extends JpaRepository<PortfolioTag, PortfolioAndTagId> {
+
+}
