--- server/src/main/java/com/konkuk/soar/repository/common/FileRepository.java
@@ -1,8 +1,0 @@
-package com.konkuk.soar.repository.common;
-
-import com.konkuk.soar.domain.common.File;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface FileRepository extends JpaRepository<File, Long> {
-
-}
