--- server/src/main/java/com/konkuk/soar/repository/member/MemberRepository.java
@@ -1,8 +1,0 @@
-package com.konkuk.soar.repository.member;
-
-import com.konkuk.soar.domain.member.Member;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface MemberRepository extends JpaRepository<Member, Long> {
-
-}
