--- server/src/main/java/com/konkuk/soar/repository/portfolio/PortfolioRepository.java
@@ -1,8 +1,0 @@
-package com.konkuk.soar.repository.portfolio;
-
-import com.konkuk.soar.domain.portfolio.Portfolio;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface PortfolioRepository extends JpaRepository<Portfolio, Long> {
-
-}
