--- server/src/main/java/com/konkuk/soar/repository/portfolio/project/ProjectRepository.java
@@ -1,8 +1,0 @@
-package com.konkuk.soar.repository.portfolio.project;
-
-import com.konkuk.soar.domain.portfolio.project.Project;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ProjectRepository extends JpaRepository<Project, Long> {
-
-}
