--- server/src/main/java/com/konkuk/soar/repository/studyhistory/StudyHistoryRepository.java
@@ -1,8 +1,0 @@
-package com.konkuk.soar.repository.studyhistory;
-
-import com.konkuk.soar.domain.studyhistory.StudyHistory;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface StudyHistoryRepository extends JpaRepository<StudyHistory, Long> {
-
-}
