+++ server/src/main/java/com/konkuk/soar/studyhistory/domain/StudyHistoryFile.java
@@ -1,0 +1,49 @@
+package com.konkuk.soar.studyhistory.domain;
+
+import com.konkuk.soar.common.domain.File;
+import com.konkuk.soar.studyhistory.domain.id.StudyHistoryAndFileId;
+import jakarta.persistence.*;
+import lombok.*;
+
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@Entity
+@Table(name = "study_history_file")
+@IdClass(StudyHistoryAndFileId.class)
+public class StudyHistoryFile {
+    @Id
+    @ManyToOne(fetch = FetchType.LAZY, cascade = CascadeType.PERSIST)
+    @JoinColumn(name = "shf_study_history_id")
+    private StudyHistory studyHistory;
+
+    @Id
+    @ManyToOne(fetch = FetchType.LAZY, cascade = CascadeType.PERSIST)
+    @JoinColumn(name = "shf_file_id")
+    private File file;
+
+    @Builder
+    public StudyHistoryFile(StudyHistory studyHistory, File file) {
+        this.studyHistory = studyHistory;
+        if (!studyHistory.getFileList().contains(this)) {
+            studyHistory.getFileList().add(this);
+        }
+
+        this.file = file;
+        if (!file.getStudyHistoryFileList().contains(this)) {
+            file.getStudyHistoryFileList().add(this);
+        }
+    }
+
+    @PreRemove
+    private void preRemove() {
+        if (studyHistory != null) {
+            studyHistory.getFileList().remove(this);
+            studyHistory = null;
+        }
+
+        if (file != null) {
+            file.getStudyHistoryFileList().remove(this);
+            file = null;
+        }
+    }
+}
