+++ server/src/main/java/com/konkuk/soar/studyhistory/repository/StudyHistoryRepository.java
@@ -1,0 +1,8 @@
+package com.konkuk.soar.studyhistory.repository;
+
+import com.konkuk.soar.studyhistory.domain.StudyHistory;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface StudyHistoryRepository extends JpaRepository<StudyHistory, Long> {
+
+}
